+++ Reflection/Changes.md
@@ -19,6 +19,7 @@
 - Added .hasSameDimensionsAs(TextDrawContext), .mapFont(...) and .mapColor(...) to TextDrawContext
 - Added .stripControlCharacters to LocalStringLike
 - Added .withShade(ColorShadeVariant) to SingleColorIcon
+- Added .forLuminosity(Double) to ColorShadeVariant
 ### Fixes
 - Alignment.position(...) now properly takes parameter bounds position into account
 - Fixed a small bug where Window content size wouldn't update correctly